Fix a compiler warning.
git-svn-id: svn://svn.tartarus.org/sgt@9775 cda61777-01e9-0310-a592-d414129be87e

+++ misc/prtargs.c
@@ -6,6 +6,7 @@ int main(int argc, char **argv)
     int i;
     printf("argc = %d\n", argc);
     for (i = 0; i < argc; i++)
-        printf("argv[%d] (%d chars) = '%s'\n", i, strlen(argv[i]), argv[i]);
+        printf("argv[%d] (%u chars) = '%s'\n", i,
+               (unsigned)strlen(argv[i]), argv[i]);
     return 0;
 }
